bug: Make tests more durable
Tests regularly fail, delaying CLI releases, because there are other
Volumes on the test account, and commands referencing a `$volume_id` are
given a list of IDs instead of just the one they expect.

This change sets `$volume_id` to a single Volume's ID, which should
allow tests to continue even if there are unexpected volumes on the
account.