bug: Remove commands with no actions from recognized commands
Related to https://github.com/linode/linode-api-docs/pull/581

It is possible that every operation in a path is skipped in the CLI via
the `x-linode-cli-skip` spec extension.  In this case, that path's
command should not be present in the CLI at all, as it would have no
actions associated to it.  The above-linked PR introduces the first
paths that are entirely skipped in this way.

This change removes all empty commands from the CLI after building the
entire command/action structure.